refactor: 추상화 제거 및 util 객체로 데이터 생성
