--- android/app/src/main/java/com/on/turip/platform/device/AppEnvironmentInfoProvider.kt
@@ -1,7 +1,0 @@
-package com.on.turip.platform.device
-
-import com.on.turip.domain.common.AppEnvironmentInfo
-
-interface AppEnvironmentInfoProvider {
-    fun getAppEnvironmentInfo(): AppEnvironmentInfo
-}
